requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
pydantic[email]>=2.7
psycopg[binary]>=3.2
sqlalchemy>=2.0
alembic>=1.13
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
pytest>=8.0
httpx>=0.24
requests>=2.31
